requirements.txt
torch==1.9.0
torchvision==0.11.2
pillow==9.0.0
black==21.12b0
flake8==4.0.1